moved code RebuildSourceCode
