+++ Assets/Sources/UseCases/Compiler/SourceCodeRebuilder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a8b49caa0cd1634799567e883b30e3e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
