+++ Assets/Tests/UseCases/Compiler/ParserTest.cs
@@ -30,6 +30,8 @@ namespace UnityLike.Test
 
             StatementNode firstParsedStatement = parser.GetParsedStatements()[0];
 
+            firstParsedStatement.LogThis();
+
             string expectedText = source;
             string parsedText = firstParsedStatement.ToPrettyString();
             Assert.AreEqual(expectedText, parsedText);
